Extract function CreateTrajectoryCarSystem.
